Added shaders and basic buffer system

+++ Atlas/src/Atlas/ImGui/ImGuiLayer.cpp
@@ -9,7 +9,7 @@
 #include "backends\imgui_impl_glfw.h"
 #include "backends\imgui_impl_opengl3.h"
 
-#include "Atlas/Application.h";
+#include "Atlas/Application.h"
 
 namespace Atlas
 {
@@ -73,7 +73,7 @@ namespace Atlas
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Window& window = Application::Get().GetWindow();
-		io.DisplaySize = ImVec2(window.GetWidth(), window.GetHeight());
+		io.DisplaySize = ImVec2((float)window.GetWidth(), (float)window.GetHeight());
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
